Minor refactoring
Preparation for #[derive(FromSubject)]
